Creating models, factories and seeders

+++ database/migrations/2019_09_09_171014_create_direcciones_table.php
@@ -17,7 +17,7 @@ class CreateDireccionesTable extends Migration
             $table->bigIncrements('id');
 
             // ID del estudiante
-            $table->unsignedTinyInteger('estudiante_id');
+            $table->unsignedBigInteger('estudiante_id');
 
             // Pais donde vive el estudiante
             $table->string('pais', 100);
